fix ingot pile being accidentally moved
